requirements/local.txt
asgiref==3.11.1
Django==5.2.12
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
pillow==12.2.0
psycopg2==2.9.11
PyJWT==2.12.1
sqlparse==0.5.5
tzdata==2024.1
